fix(tui): find search matches past row 65535
The row a match was found on was kept as a u16 while the scrollback holds up to
a million rows, so in a longer log the index wrapped around and both the
highlight and the jump to the match landed on an unrelated row.

+++ src/app/tui/search.rs
@@ -9,8 +9,9 @@ pub struct SearchResults {
     pub backward: bool,
 }
 
+/// The row and column a match starts at, counted over the whole grid including the scrollback.
 #[derive(Debug, Clone)]
-pub struct Match(pub u16, pub u16);
+pub struct Match(pub usize, pub usize);
 
 impl SearchResults {
     pub fn new(task: &str, query: String, matches: Vec<Match>, index: usize, backward: bool) -> anyhow::Result<Self> {
